# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/projects/{projectId}/issues": {
      "post": {
        "operationId": "createIssue",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIssueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a task in a project",
        "tags": ["issues"]
      }
    },
    "/api/v1/issues/{issueId}": {
      "patch": {
        "operationId": "updateIssue",
        "parameters": [
          {
            "name": "issueId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIssueDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update a task",
        "tags": ["issues"]
      }
    },
    "/api/v1/issues/{issueId}/move": {
      "post": {
        "operationId": "moveIssue",
        "parameters": [
          {
            "name": "issueId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveIssueDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponseDto"
                }
              }
            }
          }
        },
        "summary": "Move and reorder a task",
        "tags": ["issues"]
      }
    },
    "/api/v1/issues/{issueId}/archive": {
      "post": {
        "operationId": "archiveIssue",
        "parameters": [
          {
            "name": "issueId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionedIssueCommandDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponseDto"
                }
              }
            }
          }
        },
        "summary": "Archive a task",
        "tags": ["issues"]
      }
    },
    "/api/v1/issues/{issueId}/duplicate": {
      "post": {
        "operationId": "duplicateIssue",
        "parameters": [
          {
            "name": "issueId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DuplicateIssueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponseDto"
                }
              }
            }
          }
        },
        "summary": "Duplicate a task and its checklist",
        "tags": ["issues"]
      }
    },
    "/api/v1/issues/{issueId}/restore": {
      "post": {
        "operationId": "restoreIssue",
        "parameters": [
          {
            "name": "issueId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreIssueDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponseDto"
                }
              }
            }
          }
        },
        "summary": "Restore an archived task",
        "tags": ["issues"]
      }
    },
    "/api/v1/projects/{projectId}/board": {
      "get": {
        "operationId": "getBoard",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sprintId",
            "required": false,
            "in": "query",
            "description": "Return only tasks from this Active Sprint",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get the active board aggregate",
        "tags": ["boards"]
      }
    },
    "/api/v1/projects/{projectId}/columns": {
      "post": {
        "operationId": "createColumn",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateColumnDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardColumnResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a board column",
        "tags": ["boards"]
      }
    },
    "/api/v1/columns/{columnId}": {
      "patch": {
        "operationId": "updateColumn",
        "parameters": [
          {
            "name": "columnId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateColumnDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardColumnResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update a board column",
        "tags": ["boards"]
      }
    },
    "/api/v1/columns/{columnId}/move": {
      "post": {
        "operationId": "moveColumn",
        "parameters": [
          {
            "name": "columnId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveColumnDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardColumnResponseDto"
                }
              }
            }
          }
        },
        "summary": "Move and reorder a board column",
        "tags": ["boards"]
      }
    },
    "/api/v1/columns/{columnId}/clear": {
      "post": {
        "operationId": "clearColumn",
        "parameters": [
          {
            "name": "columnId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionedColumnCommandDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardColumnResponseDto"
                }
              }
            }
          }
        },
        "summary": "Archive every task in a column",
        "tags": ["boards"]
      }
    },
    "/api/v1/columns/{columnId}/archive": {
      "post": {
        "operationId": "archiveColumn",
        "parameters": [
          {
            "name": "columnId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveColumnDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardColumnResponseDto"
                }
              }
            }
          }
        },
        "summary": "Archive a board column",
        "tags": ["boards"]
      }
    },
...
  },
  "info": {
    "title": "My Kanban API",
    "description": "Private API for the personal Kanban and Scrum board",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ChecklistItemInputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "maxLength": 300
          },
          "isCompleted": {
            "type": "boolean",
            "default": false
          }
        },
        "required": ["title"]
      },
      "CreateIssueDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "columnId": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string",
            "maxLength": 50000,
            "default": ""
          },
          "type": {
            "type": "string",
            "enum": ["task", "story", "bug", "chore"],
            "default": "task"
          },
          "priority": {
            "type": "string",
            "enum": ["urgent", "high", "medium", "low", "none"],
            "default": "medium"
          },
          "labels": {
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "storyPoints": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "nullable": true
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isBlocked": {
            "type": "boolean",
            "default": false
          },
          "blockedReason": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "checklist": {
            "maxItems": 100,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChecklistItemInputDto"
            }
          },
          "beforeIssueId": {
            "type": "string",
            "format": "uuid"
          },
          "afterIssueId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["title", "columnId"]
      },
      "ChecklistItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "isCompleted": {
            "type": "boolean"
          }
        },
        "required": ["id", "title", "isCompleted"]
      },
      "IssueResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "sprintId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "columnId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["task", "story", "bug", "chore"]
          },
          "priority": {
            "type": "string",
            "enum": ["urgent", "high", "medium", "low", "none"]
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "storyPoints": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "maximum": 100
          },
          "dueDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "isBlocked": {
            "type": "boolean"
          },
          "blockedReason": {
            "type": "string",
            "nullable": true
          },
          "checklist": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChecklistItemResponseDto"
            }
          },
          "checklistIncompleteCount": {
            "type": "number",
            "minimum": 0
          },
          "completedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "version": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "projectId",
          "sprintId",
          "columnId",
          "title",
          "description",
          "type",
          "priority",
          "labels",
          "storyPoints",
          "dueDate",
          "isBlocked",
          "blockedReason",
          "checklist",
          "checklistIncompleteCount",
          "completedAt",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateIssueDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 50000,
            "default": ""
          },
          "type": {
            "type": "string",
            "enum": ["task", "story", "bug", "chore"],
            "default": "task"
          },
          "priority": {
            "type": "string",
            "enum": ["urgent", "high", "medium", "low", "none"],
            "default": "medium"
          },
          "labels": {
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "storyPoints": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "nullable": true
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isBlocked": {
            "type": "boolean",
            "default": false
          },
          "blockedReason": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "checklist": {
            "maxItems": 100,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChecklistItemInputDto"
            }
          },
          "version": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": ["version"]
      },
      "MoveIssueDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "minimum": 1
          },
          "targetColumnId": {
            "type": "string",
            "format": "uuid"
          },
          "sprintId": {
            "type": "string",
            "format": "uuid",
            "description": "Visible Sprint scope used when ordering tasks on the Scrum Board"
          },
          "beforeIssueId": {
            "type": "string",
            "format": "uuid"
          },
          "afterIssueId": {
            "type": "string",
            "format": "uuid"
          },
          "allowIncompleteChecklist": {
            "type": "boolean",
            "default": false,
            "description": "Explicitly allow moving a task with unfinished checklist items to Done"
          }
        },
        "required": ["version", "targetColumnId"]
      },
      "VersionedIssueCommandDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": ["version"]
      },
      "DuplicateIssueDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "minimum": 1
          },
          "targetColumnId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["version"]
      },
      "RestoreIssueDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "minimum": 1
          },
          "targetColumnId": {
            "type": "string",
            "format": "uuid"
          },
          "beforeIssueId": {
            "type": "string",
            "format": "uuid"
          },
          "afterIssueId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["version"]
      },
      "BoardProjectDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "enum": ["kanban", "scrum"]
          },
          "version": {
            "type": "number"
          }
        },
        "required": ["id", "name", "mode", "version"]
      },
      "BoardColumnResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": ["todo", "in_progress", "done"]
          },
          "wipLimit": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "version": {
            "type": "number"
          }
        },
        "required": ["id", "projectId", "name", "category", "wipLimit", "version"]
      },
      "BoardResponseDto": {
        "type": "object",
        "properties": {
          "project": {
            "$ref": "#/components/schemas/BoardProjectDto"
          },
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoardColumnResponseDto"
            }
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IssueResponseDto"
            }
          }
        },
        "required": ["project", "columns", "issues"]
      },
      "CreateColumnDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80
          },
          "category": {
            "type": "string",
            "enum": ["todo", "in_progress", "done"],
            "default": "todo"
          },
          "wipLimit": {
            "type": "number",
            "minimum": 1,
            "nullable": true
          },
          "beforeColumnId": {
            "type": "string",
            "format": "uuid"
          },
          "afterColumnId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["name"]
      },
      "UpdateColumnDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "minimum": 1
          },
          "name": {
            "type": "string",
            "maxLength": 80
          },
          "wipLimit": {
            "type": "number",
            "minimum": 1,
            "nullable": true
          }
        },
        "required": ["version"]
      },
      "MoveColumnDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "minimum": 1
          },
          "beforeColumnId": {
            "type": "string",
            "format": "uuid"
          },
          "afterColumnId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["version"]
      },
      "VersionedColumnCommandDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "minimum": 1
          },
          "sprintId": {
            "type": "string",
            "format": "uuid",
            "description": "Limit the command to one Active Sprint"
          }
        },
        "required": ["version"]
      },
      "ArchiveColumnDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "minimum": 1
          },
          "destinationColumnId": {
            "type": "string",
            "format": "uuid",
            "description": "Required when the archived column still contains active tasks"
          },
          "allowIncompleteChecklist": {
            "type": "boolean",
            "default": false,
            "description": "Explicitly allow moving tasks with unfinished checklist items to a Done destination"
          }
        },
        "required": ["version"]
      },
...
    }
  }
}