# packages/api-client/src/generated/model/moveIssueDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * My Kanban API
 * Private API for the personal Kanban and Scrum board
 * OpenAPI spec version: 1.0
 */

export interface MoveIssueDto {
  /** @minimum 1 */
  version: number;
  targetColumnId: string;
  /** Visible Sprint scope used when ordering tasks on the Scrum Board */
  sprintId?: string;
  beforeIssueId?: string;
  afterIssueId?: string;
  /** Explicitly allow moving a task with unfinished checklist items to Done */
  allowIncompleteChecklist?: boolean;
}